+++ packages/react-ui/components/Toast/Toast.tsx
@@ -20,6 +20,7 @@ export interface ToastState {
   notification: Nullable<string>;
   action: Nullable<Action>;
   id: number;
+  showTime: Nullable<number>;
 }
 
 export interface ToastProps extends CommonProps {
@@ -37,7 +38,7 @@ export const ToastDataTids = {
 /**
  * Показывает уведомления.
  *
- * Доступен статический метод: `Toast.push(notification, action?)`.
+ * Доступен статический метод: `Toast.push(notification, action?, showTime?)`.
  * Однако, при его использовании не работает кастомизация и могут быть проблемы
  * с перекрытием уведомления другими элементами страницы.
  *
@@ -48,8 +49,8 @@ export class Toast extends React.Component<ToastProps, ToastState> {
   public static __KONTUR_REACT_UI__ = 'Toast';
   private setRootNode!: TSetRootNode;
 
-  public static push(notification: string, action?: Action) {
-    ToastStatic.push(notification, action);
+  public static push(notification: string, action?: Nullable<Action>, showTime?: number) {
+    ToastStatic.push(notification, action, showTime);
   }
 
   public static close() {
@@ -66,6 +67,7 @@ export class Toast extends React.Component<ToastProps, ToastState> {
       notification: null,
       action: null,
       id: 0,
+      showTime: null,
     };
   }
 
@@ -85,20 +87,22 @@ export class Toast extends React.Component<ToastProps, ToastState> {
    * Показывает тост с `notification` в качестве сообщения.
    * Тост автоматически скрывается через 3 или 7 секунд,
    * в зависимости от наличия у него кнопки `action`.
+   * Время показа можно задать вручную, передав `showTime`.
    *
    * @public
    * @param {string} notification
    * @param {Action} action `action` опциональный параметр формата `{ label: string, handler: function }`
    * добавляет кнопку в виде ссылки при клике на которую вызывается переданный handler
+   * @param {number} showTime Время существования Toast в миллисекундах
    */
-  public push(notification: string, action?: Action) {
+  public push(notification: string, action?: Nullable<Action>, showTime?: number) {
     if (this.state.notification) {
       this.close();
     }
 
     safelyCall(this.props.onPush, notification, action);
 
-    this.setState(({ id }) => ({ notification, action, id: id + 1 }), this._setTimer);
+    this.setState(({ id }) => ({ notification, action, id: id + 1, showTime }), this._setTimer);
   }
 
   /**
@@ -164,9 +168,10 @@ export class Toast extends React.Component<ToastProps, ToastState> {
   private _setTimer = () => {
     this._clearTimer();
 
-    const timeOut = this.state.action ? 7 : 3;
+    let showTime = this.state.action ? 7000 : 3000;
+    showTime = this.state.showTime ?? showTime;
 
-    this._timeout = window.setTimeout(this.close, timeOut * 1000);
+    this._timeout = window.setTimeout(this.close, showTime);
   };
 
   private _refToast = (element: ToastView) => {